Wrapped code into AdminFrame object
The AdminFrame object extends the ttk.Frame (use ttk when possible, it
looks nicer) I also provided basic run code for this Frame. (its at the
bottom of the file, with root and what not) I also formatted the
AdminFrame object to standards so we can easily see what is going on.
Following the format as followed:

1. Create the widget (such as button, label, etc)
2. Apply formatting/listeners (I only applied formatting )
3. Apply to grid (I use no arguments for the grid, so it functions like
pack(), but in the future grid() will be needed for more complex layous.

Finally I added ttk.Style(), and applied the "best" looking style for
linux. Idk what it will look like elsewhere, but it looks better than
normal so we will use the theme.