Changed database interface to OOP,new rebuild.py
I changed the database interface to follow traditional OOP design
practicies. So now it is like a java class. I also added a few utility
functions to the class to assist in integration.

The rebuild.py file is a quick script that calls the new database
functions to delete and hopefully rebuild the database file.