Moved rebuild back, fixed database creation
Moving rebuild messed things up, because when the program executes it
starts from the main.py, so the database WILL be in the local root
folder without any customization.
So it only makes sense to leave the rebuild.py in the root folder as
well.

I also updated the databaseinterface to save the connection to the
database, this added a small issue of creating the database agian, but
this was also fixed.