Added new column to prinout, and bugfixes
Fixed bug where actuall ?? would be placed into the database instead of
null. This would be done during the check phase of the database
functions. I removed this code until I figure out the date issue.

The new column added to the printout will show the HRs and minutes
between clocking in and clocking out. Need to test this system first.