Added date calulation for row.
Automatically calculates date IF it is given. If it is null, there will
be ???? marks made.

Also fixed PATH during database printout. So it prints the actual path
of the current working directory. The text format is all messed up, but
whatever haha.

Finally fixed a minor format issue with the headers in the printout. The
Last column, Valid wasn't formatted correctly, this should be fixed.