Fixed database bugs with reporting
Had to redesign issues with the database interface not keeping track of
the database integrity in all cases.

The ioHandler was fixed to display calculated hours only if its a valid
entry. The automatic clockout time automatically shows, but with a date
that is a day later ( at the least). This isn't the same as a user that
is currently clocked in (their clockout time will be ????)

Also fixed issue where the app would say successful login, even if the
user is already clocked in.