Fixed issue with 2nd print to file not working
This was due to changing the directory for no reason. As V said there is
a reason for everything! The lines removed changed the CWD to the
EmployeeHome file, thus making and subsequent calls invalid.

I also moved around how I handle the database, this should result in
slightly quicker more effective code.